web_chat/requirements.txt
django
channels==4.2.0
redis==5.2.1
channels-redis==4.2.1
psycopg2==2.9.10
djangorestframework==3.15.2
requests
django-cors-headers
loguru
daphne
boto3